refactor: move template date from util.js to data.js
